Updated audio_gen.c to generate float
audio_gen.c now produces its results in floats instead of in fixed point, this should give a better result when the signal later is processed.

As a result of the above, many functions of audio_gen.c now uses floating point instead of fixed point.

To reduce the number of calculations, the triangle/saw slope and sine rad/sample is now calculated only once each loop.

mono_to_stereo() now handles float to uint conversion, this function will at a later point handle all hardware specific conversion.

Added new waves: inverted triangle and dobule saw.

+++ Inc/audio_interface.h
@@ -16,7 +16,7 @@
 
 void init_audio_output(void);
 //void add_to_audio_buffer(int16_t *audio_buffer, uint32_t buf_len);
-void mono_to_stereo (uint16_t *stereo_samples, uint16_t *mono_samples, uint16_t n_sample_mono);
+void mono_to_stereo (uint16_t *stereo_samples, float *mono_samples, uint16_t n_sample_mono);
 void send_audio(void);
